minor clean up of lattice examples
